+++ src/Util/ITimer.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+
+namespace FSM
+{
+	public interface ITimer
+	{
+		float Elapsed {
+			get;
+		}
+
+		void Reset();
+	}
+}